Rewrote `JavaKotlinPojoGenerator` to be a little simpler
The old approach was to have a `JavaGenerator` with a supplementary `KotlinGenerator` that overrides pojo/interface functions (along with a few other implementation detail functions). This worked but felt a little hard to follow (as it often feels with overriding concrete functions).

The new approach simply takes the `JavaGenerator` and `KotlinGenerator`, lets them both run to finish and then deletes any pojo/interface java files, and any non-pojo/non-interface kotlin files. This leaves you with the same result with a process that's easier to hold in your head.

This approach is similar to approach suggested in https://github.com/jOOQ/jOOQ/issues/13338 (but instead of configuring in Gradle/Maven, it's encapsulated as a Generator that can be published to Maven and plugged in as a one-line in the Jooq Gradle Plugin!).